feat(branding) - use executable-based values
In an effort to make rebranding the CLI app a bit easier, this
changes all parsers, cmds, and settings to use a value based on
the executable name rather than the hard-coded "deis" values.

As things stand right now, the executable name built by `make` is
still `deis`, so all online help via `deis help` will still refer
to `deis`, `DEIS_PROFILE`, and `~/.deis/config.json`, and the
default remote name is still `deis`.

If on windows (where the executable is `deis.exe`), much of this is
the same, except that the commands suggested in the online help
refer to `deis.exe` rather than just `deis`.

Renaming the executable to anything modifies all of the above to
use the new exectuable name as the basis for the information
presented.